refactor: rename components and change structure
