+++ src/components/BlockImages/BlockImages.classname.js
@@ -1,0 +1,5 @@
+import { cn } from '@bem-react/classname'
+
+const cnBlockImages = cn('BlockImages')
+
+export default cnBlockImages
