+++ src/components/BlockImages/BlockImagesImage/BlockImagesImage.css
@@ -1,12 +1,12 @@
-.wrapper-image {
+.BlockImages-Image {
   flex: 0 0 calc(25% - 16px);
 }
 
-.wrapper-image:hover {
+.BlockImages-Image:hover {
   cursor: pointer;
 }
 
-.image {
+.BlockImages-ImageElemnt {
   width: 100%;
   height: 234px;
   object-fit: cover;
