+++ src/components/SearchBar/SearchBar.classname.js
@@ -1,0 +1,5 @@
+import { cn } from '@bem-react/classname'
+
+const cnSearchBar = cn('SearchBar')
+
+export default cnSearchBar
