+++ src/components/SearchBar/SearchBarError/SearchBarError.jsx
@@ -1,0 +1,9 @@
+import cnSearchBar from '../SearchBar.classname'
+
+import './SearchBarError.css'
+
+const SearchBarError = ({ error }) => (
+  <div className={cnSearchBar('Error')}>{error}</div>
+)
+
+export default SearchBarError
