+++ src/components/SearchBar/hooks/useGetValue.jsx
