--- src/components/SearchImagesBack/BlockImages/BlockImagesImage/BlockImagesImage.jsx
@@ -1,14 +1,0 @@
-import './BlockImagesImage.css'
-
-const BlockImagesImage = ({ dataImg }) => {
-  const { urls } = dataImg
-  const description = dataImg.alt_description
-
-  return (
-    <div className="wrapper-image">
-      <img src={urls.regular} alt={description} className="image" />
-    </div>
-  )
-}
-
-export default BlockImagesImage
